# static/swagger.yml
openapi: 3.0.0
info:
  title: API de Gestión de Dispositivos Wake On LAN (WoL)
  description: API de Gestión de Dispositivos Wake On LAN (WoL) y otros del hogar
  version: 1.0.0
paths:
...
  /add_device:
    get:
      summary: Renderiza el formulario para agregar un nuevo dispositivo
      tags:
        - Dispositivos
      responses:
        "200":
          description: Página de añadir dispositivo
        "302":
          description: Redirección al login si no está autenticado
    post:
      summary: Crear un nuevo dispositivo
      tags:
        - Dispositivos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nombre del dispositivo
                ip_address:
                  type: string
                  description: Dirección IP del dispositivo
                mac_address:
                  type: string
                  description: Dirección MAC del dispositivo
                device_type:
                  type: string
                  description: Tipo de dispositivo
              required:
                - name
                - device_type
      responses:
        "302":
          description: Redirección al dashboard si es exitoso
        "400":
          description: Datos inválidos
  /remove_device:
    delete:
      summary: Eliminar un dispositivo
      tags:
        - Dispositivos
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          description: Nombre del dispositivo a eliminar
        - name: tag
          in: query
          required: true
          schema:
            type: string
          description: Identificador único del dispositivo
      responses:
        "204":
          description: Dispositivo eliminado con éxito
        "400":
          description: Parámetros faltantes
  /update_device_name:
    put:
      summary: Update the name of an existing device
      description: Update a device name from the system using its name and tag identifiers
      tags:
        - Dispositivos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the device
                tag:
                  type: string
                  description: Unique tag identifier of the device
              required:
                - name
                - tag
      responses:
        "200":
          description: Device's name successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: "True"
        "400":
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      error:
                        type: string
                        description: "No se proporcionaron datos JSON"
                  - type: object
                    properties:
                      error_tag_name:
                        type: string
                        description: "Name or Tag missing"
...